requirements_dev.txt
pip==9.0.3
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
Sphinx==1.7.1
twine==1.10.0
nbconvert==5.3.1
requests>=2.8,<3.0
flask-restful==0.3.6
jupyter_enterprise_gateway>=1.0.0
